% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesMixture.R
\name{bayesMixture}
\alias{bayesMixture}
\title{bayesMixture}
\usage{
bayesMixture(
  n = 20,
  y = 12,
  p = 0.3,
  alpha1 = 2,
  beta1 = 3,
  alpha2 = 4,
  beta2 = 5,
  a = 0.025
)
}
\arguments{
\item{n}{number of trials in the binomial experiment}

\item{y}{number of successes}

\item{p}{the mixing probability, the weight of the first beta distribution}

\item{alpha1}{alpha parameter of the first prior Beta distribution}

\item{beta1}{beta parameter of the first prior Beta distribution}

\item{alpha2}{alpha parameter of the second prior Beta distribution}

\item{beta2}{beta parameter of the second prior Beta distribution}

\item{a}{significance level}
}
\value{
named list containing Bayesian point and interval estimates
}
\description{
This function produces a plot of the mixture prior, and a plot of the posterior.
It also print a named list of Bayesian point and interval estimates of theta for a given tail area of a
}
\examples{
bayesMixture(n=20, y=12, p=0.3, alpha1=2, beta1=3, alpha2=4, beta2=5, a=0.025)
}
