% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesPosterior.R
\name{bayesPosterior}
\alias{bayesPosterior}
\title{bayesPosterior}
\usage{
bayesPosterior(n = 10, y = 4, alpha = 1, beta = 1, a = 0.025)
}
\arguments{
\item{n}{number of trials in the binomial experiment}

\item{y}{number of successes}

\item{alpha}{alpha parameter of the prior Beta distribution}

\item{beta}{beta parameter of the prior Beta distribution.}

\item{a}{significance level}
}
\value{
named list containing Bayesian point and interval estimates
}
\description{
The function produces a plot that has all three densities on the one axes, p(theta), f(y|theta),p(theta|y)
It prints a named list containing Bayesian point and interval estimates (intervals made for a given tail area of a)
}
\examples{
bayesPosterior (n = 10, y = 4, alpha = 1, beta = 1, a = 0.025)
}
