% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertable.R
\name{clustertable}
\alias{clustertable}
\title{clustertable}
\usage{
clustertable(df, original = ddt, name = "SPECIES", k = NULL)
}
\arguments{
\item{df}{table of dataset, obtaib from the original, that need to do the cluster}

\item{original}{table of dataset, only with Quantitative data}

\item{name}{name of the column, going to do the cluster}

\item{k}{number of cluster}
}
\value{
a table and the plot of cluster
}
\description{
clustertable
}
\examples{
df <- data.frame(Length=ddt$LENGTH, Weight = ddt$WEIGHT); clustertable(df, original=ddt, name = "SPECIES")
}
