% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.R
\name{corMatrix}
\alias{corMatrix}
\title{the sample correlation matrix}
\usage{
corMatrix(x)
}
\arguments{
\item{x}{a set of quantitative date with multiple vector}
}
\value{
the matrix of the sample correlation
}
\description{
the sample correlation matrix
}
