% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatrix.R
\name{covMatrix}
\alias{covMatrix}
\title{the sample (biased) covariance matrix}
\usage{
covMatrix(x)
}
\arguments{
\item{x}{a set of quantitative date with multiple vector}
}
\value{
the matrix of the sample covariance
}
\description{
the sample (biased) covariance matrix
}
\examples{
data <- data.frame(x = rnorm(10), y = rnorm(10)); covMatrix(data)
}
