% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean.R
\name{mymean}
\alias{mymean}
\title{mymean}
\usage{
mymean(x)
}
\arguments{
\item{x}{a set of quantitative date with multiple vector}
}
\value{
the mean of each vector
}
\description{
mymean
}
\examples{
data <- data.frame(x = rnorm(10), y = rnorm(10)); mymean(data)
}
