% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymean1.R
\name{mymean1}
\alias{mymean1}
\title{mymean1}
\usage{
mymean1(x)
}
\arguments{
\item{x}{the vector of the data}
}
\value{
the mean of the vector
}
\description{
mymean1
}
\examples{
 data <- data.frame(x = rnorm(10), y = rnorm(10)); apply(data, 2, mymean1)
}
