% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{Basic square funtion}
\usage{
mysquare(x)
}
\arguments{
\item{x}{a vector of quantitative date}
}
\value{
the square of the input vector
}
\description{
Basic square funtion
}
