% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T1-2.R
\docType{data}
\name{T1_2}
\alias{T1_2}
\title{T1-2 data}
\format{
A data frame with 41 rows and 3 columns
\describe{
  \item{V1}{Density}
  \item{V2}{Machine direction}
  \item{V3}{Cross direction}
  ...
}
}
\source{
text book
}
\usage{
T1_2
}
\description{
41 lines of Paper-Quality Measurements
}
\keyword{datasets}
