% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorMean.R
\name{vectorMean}
\alias{vectorMean}
\title{the sample mean vector}
\usage{
vectorMean(x)
}
\arguments{
\item{x}{a set of quantitative date with multiple vector}
}
\value{
the mean of each vector
}
\description{
the sample mean vector
}
\examples{
data <- data.frame(x = rnorm(10), y = rnorm(10)); vectorMean(data)
}
